First draft of loose Docker guidelines/pointers.
Goal here is to help get folks started on using Docker for development.
Unfortunately, we don't have a cookie-cutter template for this sort of thing,
so it's less comprehensive than I'd like. It does try to add some fencing
(e.g. use docker-compose, here are some options for one-off commands, etc.)
and links to several projects.