Fix security related URLs
The permalink was missing the trailing slash for the CSP page
which resulted in the URL having a .html file extension, while
the rest of the documents do not. I updated the permalink to
account for this and added both a security main page as well
as data for the security side navigation.