No index.md nesting, fix Federalist deploy
Without a trailing slash on `:path` in `_config.yml`, `jekyl serve` will
happily (but incorrectly) serve up '/frontend`, etc. but it only
generates `_site/frontend.html`, etc. This results in a working local
site, but a broken Federalist deploy.

A good consequence is that we can rename the nested index.md files,
moving them up a level and giving them a meaningful filename.