Initial pass of changing 18F to TTS
An overdue swap of 18F for TTS in the guide content, since the
guide is a product of the TTS Engineering Practice Guild. It is
a little imprecise when describing some trends or practices, as
much of the original content was only done with 18F in mind, but
it is good enough as a start.